console command uses new font
tune occlusion culling
